+++ src/main/java/org/codequistify/master/application/player/service/PlayerProfileService.java
@@ -3,7 +3,7 @@ package org.codequistify.master.application.player.service;
 import lombok.RequiredArgsConstructor;
 import org.codequistify.master.application.player.dto.PlayerProfile;
 import org.codequistify.master.application.player.dto.PlayerStageProgressResponse;
-import org.codequistify.master.core.domain.player.domain.PlayerRoleType;
+import org.codequistify.master.core.domain.player.model.PlayerRoleType;
 import org.codequistify.master.core.domain.player.model.Player;
 import org.codequistify.master.core.domain.player.service.PlayerRolesChecker;
 import org.codequistify.master.core.domain.stage.dto.HeatMapDataPoint;
