+++ src/main/java/org/codequistify/master/application/player/service/PlayerQueryService.java
@@ -2,10 +2,10 @@ package org.codequistify.master.application.player.service;
 
 import lombok.RequiredArgsConstructor;
 import org.codequistify.master.application.exception.ApplicationException;
-import org.codequistify.master.core.domain.player.domain.OAuthType;
+import org.codequistify.master.core.domain.player.model.OAuthType;
 import org.codequistify.master.core.domain.player.model.Player;
 import org.codequistify.master.core.domain.player.model.PolId;
-import org.codequistify.master.core.domain.port.PlayerReader;
+import org.codequistify.master.core.domain.player.port.PlayerReader;
 import org.codequistify.master.global.exception.ErrorCode;
 import org.codequistify.master.infrastructure.player.converter.PlayerConverter;
 import org.codequistify.master.infrastructure.player.repository.PlayerRepository;
