+++ src/main/java/org/codequistify/master/core/domain/authentication/service/AuthenticationService.java
@@ -6,7 +6,7 @@ import lombok.RequiredArgsConstructor;
 import org.codequistify.master.core.domain.authentication.dto.LogInRequest;
 import org.codequistify.master.core.domain.authentication.dto.SignUpRequest;
 import org.codequistify.master.core.domain.player.converter.PlayerConverter;
-import org.codequistify.master.core.domain.player.domain.OAuthType;
+import org.codequistify.master.core.domain.player.model.OAuthType;
 import org.codequistify.master.core.domain.player.model.Player;
 import org.codequistify.master.application.player.dto.PlayerProfile;
 import org.codequistify.master.core.domain.player.service.PlayerDetailsService;
