+++ src/main/java/org/codequistify/master/core/domain/authentication/service/impl/GithubSocialSignService.java
@@ -6,7 +6,7 @@ import org.codequistify.master.core.domain.authentication.vo.OAuthData;
 import org.codequistify.master.core.domain.authentication.vo.OAuthResource;
 import org.codequistify.master.core.domain.authentication.vo.OAuthToken;
 import org.codequistify.master.core.domain.player.converter.PlayerConverter;
-import org.codequistify.master.core.domain.player.domain.OAuthType;
+import org.codequistify.master.core.domain.player.model.OAuthType;
 import org.codequistify.master.core.domain.player.model.Player;
 import org.codequistify.master.application.player.dto.PlayerProfile;
 import org.codequistify.master.core.domain.player.service.PlayerDetailsService;
