+++ src/main/java/org/codequistify/master/core/domain/player/converter/PlayerConverter.java
@@ -1,13 +1,14 @@
 package org.codequistify.master.core.domain.player.converter;
 
 import org.codequistify.master.core.domain.authentication.dto.SignUpRequest;
-import org.codequistify.master.core.domain.player.domain.OAuthType;
+import org.codequistify.master.core.domain.player.model.OAuthType;
 import org.codequistify.master.core.domain.player.model.Player;
 import org.codequistify.master.application.player.dto.PlayerProfile;
 import org.springframework.stereotype.Component;
 
 @Component
-public class PlayerConverter {
+public class
+layerConverter {
     public PlayerProfile convert(Player player) {
         return new PlayerProfile(
                 player.getUid(),
