+++ src/main/java/org/codequistify/master/core/domain/player/port/PlayerReader.java
@@ -1,4 +1,4 @@
-package org.codequistify.master.core.domain.port;
+package org.codequistify.master.core.domain.player.port;
 
 import org.codequistify.master.core.domain.player.model.Player;
 import org.codequistify.master.core.domain.player.model.PolId;
